decoder: Rename `mod logger` to `log`
