+++ decoder/src/lib.rs
@@ -6,4 +6,4 @@ include!(concat!(env!("OUT_DIR"), "/version.rs"));
 
 pub mod decoder;
 pub mod elf2table;
-pub mod logger;
+pub mod log;
