+++ decoder/src/log.rs
